solutions/mod-07/requirements/behave-postgres.txt
behave
django
psycopg[binary,pool]
testcontainers[postgres]
pytest
pytest-env
